refactor: Enable react-hot-loader
